+++ .pre-commit-config.yaml
@@ -13,7 +13,7 @@ repos:
       - id: black
         language_version: python3
         args:
-          - "--exclude=alembic/**"
+          - "--exclude=alembic/"
   - repo: https://github.com/pycqa/flake8
     rev: "6.1.0"
     hooks:
