+++ cruds/comments/__init__.py
@@ -1,6 +1,6 @@
 from .comment import (
     create_comment,
+    delete_by_comment_id,
     get_comments_by_work_id,
     get_reply_comments_by_comment_id,
-    delete_by_comment_id,
 )
