+++ cruds/favorite/__init__.py
@@ -1,0 +1,6 @@
+from .favorite import (
+    delete_favorite_by_id,
+    get_favorite_by_user_id,
+    get_favorite_by_work_id,
+    set_favorite,
+)
