+++ routers/comments/__init__.py
@@ -1,6 +1,1 @@
-from .comment import (
-    post_comment,
-    get_comments,
-    get_reply_comments,
-    delete_by_comment_id,
-)
+from .comments import comment_router
