+++ routers/favorites/__init__.py
@@ -1,0 +1,1 @@
+from .favorites import favorite_router, user_favorite_router
