+++ routers/tags/__init__.py
@@ -1,0 +1,1 @@
+from .tags import tag_router
