+++ schemas/url_info.py
@@ -4,7 +4,7 @@ from datetime import datetime
 from pydantic import BaseModel
 from pydantic.class_validators import validator
 
-from db.models import UrlType
+from db import UrlType
 
 from .user import User
