+++ schemas/work.py
@@ -4,7 +4,7 @@ from typing import List, Optional
 from pydantic import BaseModel
 from pydantic.class_validators import validator
 
-from db.models import Visibility
+from db import Visibility
 from schemas.tag import GetTag
 from schemas.url_info import BaseUrlInfo, UrlInfo
